move [b-collapse] from Content to Directive
