+++ src/pages/Demo/Directive/b-modal-demo.vue
@@ -1,5 +1,7 @@
 <template>
   <div>
+    <h2>b-modal(v-modal)</h2>
+    <hr>
     <b-modal
       id="exampleModal"
       title="title text"
